+++ lib/prouterd/iface.rb
@@ -12,6 +12,7 @@ module Prouterd
     autoload :LlmCaller,      File.expand_path("iface/llm_caller",      __dir__)
     autoload :LlmSubprocess,  File.expand_path("iface/llm_subprocess",  __dir__)
     autoload :PostgresCaller, File.expand_path("iface/postgres_caller", __dir__)
+    autoload :LocalRepoCaller, File.expand_path("iface/local_repo_caller", __dir__)
   end
 end
 
@@ -25,3 +26,4 @@ require_relative "iface/plugins/llm"
 require_relative "iface/plugins/postgres"
 require_relative "iface/plugins/docker"
 require_relative "iface/plugins/shell"
+require_relative "iface/plugins/local_repo"